1.add redis support; 2.add game/gate project as sample; 3.add x64 support; 4.update library configure; 5.update eclipse CDT configure on centos7; 6.update VLD to 64-bit, remove 32-bit dll; 7.update eclipse project mklink; 8.add some unit test suits/cases.

+++ Source/Common/Network/DownloadMgr.cpp
@@ -36,7 +36,7 @@ namespace Network
 	DownloadTask* DownloadMgr::addTask(std::string strURL, const std::string& strLocal, unsigned int size)
 	{
 		// 解析URL的协议类型，调用对应的协议下载
-		int pos = strURL.find_first_of(":");
+		int pos = (int)strURL.find_first_of(":");
 		if (pos == -1)
 			return 0;
 		std::string strProtocol = strURL.substr(0, pos);
